+++ crates/frontend/src/lib.rs
@@ -3,10 +3,10 @@
 mod api;
 pub mod app;
 pub mod components;
-pub mod form_types;
 pub mod hoc;
 pub mod pages;
 pub mod providers;
+pub mod schema;
 pub mod types;
 mod utils;
 use cfg_if::cfg_if;
